requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
sentence-transformers
faiss-cpu
streamlit
httpx